Componentizing the Room View
