+++ components/room-view/RoomTableItemMemberEmpty.js
@@ -1,0 +1,7 @@
+import { Box } from '@chakra-ui/react';
+
+function RoomTableItemMemberEmpty() {
+	return <Box h={6} w={6} rounded="md" borderColor="outline.500" borderWidth={1} borderStyle="dotted" />;
+}
+
+export default RoomTableItemMemberEmpty;
