+++ components/room-view/index.js
@@ -1,2 +1,9 @@
+export { default as RoomTableItemActionButtons } from './RoomTableItemActionButtons';
+export { default as RoomTableItemInfo } from './RoomTableItemInfo';
+export { default as RoomTableItemMemberAvatar } from './RoomTableItemMemberAvatar';
+export { default as RoomTableItemMemberEmpty } from './RoomTableItemMemberEmpty';
+export { default as RoomTableItemMembers } from './RoomTableItemMembers';
+export { default as RoomTableItemMeta } from './RoomTableItemMeta';
+export { default as RoomTableItem } from './RoomTableItem';
 export { default as RoomView } from './RoomView';
 export { default } from './RoomView';
